ref: improve code structure
