+++ contracts/src/utils/cryptography/eip712.rs
@@ -12,11 +12,9 @@
 
 use alloc::{borrow::ToOwned, string::String, vec::Vec};
 
-use alloy_primitives::{keccak256, Address, FixedBytes, U256};
+use alloy_primitives::{keccak256, Address, B256, U256};
 use alloy_sol_types::{sol, SolType};
 use stylus_sdk::{block, contract};
-
-use crate::utils::cryptography::message_hash_utils::to_typed_data_hash;
 
 /// keccak256("EIP712Domain(string name,string version,uint256 chainId,address
 /// verifyingContract)")
@@ -32,10 +30,34 @@ pub const FIELDS: [u8; 1] = [0x0f];
 /// Salt for the domain separator.
 pub const SALT: [u8; 32] = [0u8; 32];
 
+/// Prefix for ERC-191 version with `0x01`.
+pub const TYPED_DATA_PREFIX: [u8; 2] = [0x19, 0x01];
+
 /// Tuple for the domain separator.
 pub type DomainSeparatorTuple = sol! {
     tuple(bytes32, bytes32, bytes32, uint256, address)
 };
+
+/// Returns the keccak256 digest of an EIP-712 typed data (ERC-191 version
+/// `0x01`).
+///
+/// The digest is calculated from a `domain_separator` and a `struct_hash`, by
+/// prefixing them with `[TYPED_DATA_PREFIX]` and hashing the result. It
+/// corresponds to the hash signed by the [eth_signTypedData] JSON-RPC method as
+/// part of EIP-712.
+///
+/// [eth_signTypedData]: https://eips.ethereum.org/EIPS/eip-712
+#[must_use]
+pub fn to_typed_data_hash(
+    domain_separator: &[u8; 32],
+    struct_hash: &[u8; 32],
+) -> B256 {
+    let mut preimage = [0u8; 66];
+    preimage[..2].copy_from_slice(&TYPED_DATA_PREFIX);
+    preimage[2..34].copy_from_slice(domain_separator);
+    preimage[34..].copy_from_slice(struct_hash);
+    keccak256(preimage)
+}
 
 /// EIP-712 Contract interface.
 pub trait IEIP712 {
@@ -87,13 +109,13 @@ pub trait IEIP712 {
     /// # Arguments
     ///
     /// * `&self` - Read access to the contract's state.
-    fn domain_separator_v4(&self) -> FixedBytes<32> {
+    fn domain_separator_v4(&self) -> B256 {
         let encoded = DomainSeparatorTuple::encode(&(
             TYPE_HASH,
             Self::HASHED_NAME,
             Self::HASHED_VERSION,
             Self::chain_id(),
-            Self::contract_address().into(),
+            Self::contract_address(),
         ));
 
         keccak256(encoded)
@@ -107,10 +129,7 @@ pub trait IEIP712 {
     /// # Arguments
     ///
     /// * `&self` - Read access to the contract's state.
-    fn hash_typed_data_v4(
-        &self,
-        struct_hash: FixedBytes<32>,
-    ) -> FixedBytes<32> {
+    fn hash_typed_data_v4(&self, struct_hash: B256) -> B256 {
         let domain_separator = self.domain_separator_v4();
         to_typed_data_hash(&domain_separator, &struct_hash)
     }
@@ -120,7 +139,7 @@ pub trait IEIP712 {
 mod tests {
     use alloy_primitives::{address, uint, Address, U256};
 
-    use super::{FIELDS, IEIP712, SALT};
+    use super::{to_typed_data_hash, FIELDS, IEIP712, SALT};
 
     const CHAIN_ID: U256 = uint!(42161_U256);
 
@@ -154,5 +173,25 @@ mod tests {
         assert_eq!(CONTRACT_ADDRESS, domain.4);
         assert_eq!(SALT, domain.5);
         assert_eq!(Vec::<U256>::new(), domain.6);
+    }
+
+    #[test]
+    fn test_to_typed_data_hash() {
+        // TYPE_HASH
+        let domain_separator = b256!(
+            "8b73c3c69bb8fe3d512ecc4cf759cc79239f7b179b0ffacaa9a75d522b39400f"
+        );
+        // bytes32("stylus");
+        let struct_hash = b256!(
+            "7379746c75730000000000000000000000000000000000000000000000000000"
+        );
+        let expected = b256!(
+            "cefc47137f8165d8270433dd62e395f5672966b83a113a7bb7b2805730a2197e"
+        );
+
+        assert_eq!(
+            expected,
+            to_typed_data_hash(&domain_separator, &struct_hash),
+        );
     }
 }
