+++ contracts/src/utils/cryptography/mod.rs
@@ -1,4 +1,3 @@
 //! Smart Contracts with cryptography.
 pub mod ecdsa;
 pub mod eip712;
-pub mod message_hash_utils;
